send_to_kafka/requirements.txt
quixstreams==3.2.1
python-dotenv
requests